requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
fastai==2.8.0
kornia==0.8.0
nibabel==5.3.2
albumentations==2.0.5
opencv-python==4.11.0.86
numpy==2.0.2
scipy==1.14.1
pandas==2.2.2
matplotlib==3.10.0
scikit-learn==1.6.1
tqdm==4.67.1
einops==0.8.1
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
fastai==2.8.0
kornia==0.8.0
albumentations==2.0.5
opencv-python==4.11.0.86
Pillow>=9.5.0
nibabel==5.3.2
imageio>=2.33.1
numpy==2.0.2
scipy==1.14.1
pandas==2.2.2
scikit-learn==1.6.1
matplotlib==3.10.0
tqdm==4.67.1
einops==0.8.1
ipywidgets>=7.7.1